requirements.txt
Django>=4.2
djangorestframework
drf-yasg
Faker
django-cors-headers
